#574: the wedge breadcrumb loses its top bits, so section 773 reads as 5
g_436_loop_section recorded a stuck vCPU's position in a uint8_t, but the
section IDs are deliberately three digits so that a phase groups under one
leading digit (7xx). Everything from 761 up was truncated, and three of the
results aliased real sections: 773->5, 774->6, 775->7. 768 stored as 0,
which is what a breadcrumb that was never written looks like.

This variable exists only to say WHERE a wedged vCPU was, so a wrong value
sends the reader to the wrong code. Widen the storage to uint16_t rather
than shrink the IDs, which would throw away the grouping. Both readers
(the device-lock hold diagnostic and the AP wedge report) already hold it
in an `unsigned`, so they widen with it, and the allocation sizes off the
pointer so it scales on its own.

clang had been reporting this 12 times as -Wconstant-conversion in a build
whose warnings nobody read. Promote that one warning to an error: a silent
numeric truncation has no benign form here. Verified the guard rejects the
ticket's own case, `uint8_t = 773`, with "changes value from 773 to 5".

Clean build: 12 constant-conversion warnings -> 0. Unit tests unchanged.

+++ boot/main.c
@@ -2900,7 +2900,7 @@ static uint32_t *g_ap_timer_reload;
 static uint32_t (*g_vcpu_timer_reload)[HYPE_MAX_VCPUS_PER_VM];
 /* #436: AP run-loop section breadcrumb. The AP sets these; the BSP prints them.
  * Pinpoints WHERE the loop stops when the AP freezes (deterministic ~127k iters). */
-volatile uint8_t *g_436_loop_section;
+volatile uint16_t *g_436_loop_section;
 volatile uint64_t *g_436_loop_iter;
 volatile uint64_t *g_436_last_reason;
 volatile uint64_t *g_436_last_rip;
@@ -10169,7 +10169,7 @@ wait_for_sipi:
                                  (unsigned long long)g_bsp_probe_reason[vm_idx],
                                  (unsigned long long)g_bsp_probe_rip[vm_idx],
                                  bp_guest ? "IN-GUEST" : "IN-HOST", bp_ms,
-                                 g_436_loop_section[vm_idx]);
+                                 (unsigned)g_436_loop_section[vm_idx]);
             }
         }
     }
@@ -24027,7 +24027,7 @@ EFI_STATUS EFIAPI efi_main(EFI_HANDLE ImageHandle, EFI_SYSTEM_TABLE *SystemTable
                                             (unsigned long long)g_bsp_probe_rip[pv],
                                             p_guest ? "IN-GUEST" : "IN-HOST",
                                             (p_ref != 0ull) ? (pnow - p_ref) * 1000ull / hz : 0ull,
-                                            g_436_loop_section[pv]);
+                                            (unsigned)g_436_loop_section[pv]);
                                     }
                                 }
                                 /* #92 diag: name the CPUID leaves / MSRs a spinning guest hammers.
